12/07/22

+++ Microservice/OrderService/src/main/java/com/example/demo/service/OrderService.java
@@ -2,15 +2,18 @@ package com.example.demo.service;
 
 import com.example.demo.client.ItemFeignClient;
 import com.example.demo.dto.ItemDTO;
+import com.example.demo.dto.OrderDetailDTO;
+import com.example.demo.dto.ResponseObjectEntity;
 import com.example.demo.entity.Order;
 import com.example.demo.entity.OrderDetail;
-import com.example.demo.repository.OrderDetailRepository;
 import com.example.demo.repository.OrderRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,15 +23,41 @@ public class OrderService {
     private OrderRepository orderRepository;
 
     @Autowired
-    private OrderDetailRepository orderDetailRepository;
-
+    private OrderDetailService orderDetailService;
     @Autowired
     private ItemFeignClient itemFeignClient;
 
     @Autowired
     private RabbitTemplate rabbitTemplate;
 
+    public Order saveOrder(List<OrderDetail> orderDetails, Long storeId) {
+        Order order = new Order();
+        orderRepository.save(order);
+        order.setOrderDate(LocalDate.now());
+        order.setStoreId(storeId);
+        double totalPrice = 0;
+        for (int i = 0; i < orderDetails.size(); i++) {
+            OrderDetail orderDetail = orderDetails.get(i);
+            orderDetail.setId(0);
+            ItemDTO item = itemFeignClient.getItemById(orderDetail.getItemId()).getBody();
+            if (item == null) {
+                orderDetails.remove(i--);
+                continue;
+            }
+            totalPrice += item.getPrice() * orderDetail.getItemQuantity();
+            orderDetail.setOrder(order);
+        }
+        if (totalPrice >= 10000) {
+            totalPrice = totalPrice * 0.95;
+        }
+        order.setTotalPrice(totalPrice);
+        order.setOrderDetails(orderDetails);
+        return orderRepository.save(order);
+    }
+
+
 //    public Order updateOrderById(Order order, long id) {
+//
 //        orderRepository.findById(id);
 //        order.setStore(order.getStore());
 //        order.setDeliveryNotes(order.getDeliveryNotes());
@@ -38,7 +67,6 @@ public class OrderService {
 //        return orderRepository.save(order);
 //    }
 
-
     public List<Order> getAllOrder() {
         return orderRepository.findAll();
     }
@@ -46,24 +74,6 @@ public class OrderService {
     public List<Order> getAllOrderByStoreId(long id) {
         return orderRepository.myStoreQueryByStoreId(id);
     }
-
-//    public Order saveOrder(Order order) {
-//        List<OrderDetail> orderDetails = orderDetailRepository.orderDetailsByOrderId(order.getId());
-//        List<ItemDTO> allItem = itemFeignClient.getAllItem();
-//        orderRepository.save(order);
-//        double totalPrice = 0;
-//        for (OrderDetail orderDetail : orderDetails) {
-//            for (ItemDTO itemDTO : allItem) {
-//                if (orderDetail.getItemId() == itemDTO.getId()) {
-//                    totalPrice+=orderDetail.getItemQuantity()*itemDTO.getPrice();
-//                }
-//            }
-//        }
-//        order.setOrderDate(LocalDate.now());
-//        System.out.println(order.getOrderDate());
-//        order.setTotalPrice(totalPrice);
-//        return orderRepository.save(order);
-//    }
 
     public Order getOrderById(long id) {
         Optional<Order> order = orderRepository.findById(id);
@@ -73,16 +83,44 @@ public class OrderService {
         return null;
     }
 
-    public void deleteOrderById(long id) {
-        orderRepository.deleteById(id);
+    public boolean deleteOrderByIdFromStore(long id, long storeId) {
+        Optional<Order> order = orderRepository.findById(id);
+        if (order.isPresent()) {
+            if (order.get().getOrderDate().equals(LocalDate.now())) {
+                orderRepository.deleteById(id);
+                return true;
+            }
+            return false;
+        }
+        return false;
     }
 
     public List<Order> getOrderByStoreId(long id) {
         return orderRepository.myStoreQueryByStoreId(id);
     }
 
-    public void getOrderByIdDemorabbit(Long id){
-        System.out.println(orderRepository.findById(id).get());
-        rabbitTemplate.convertAndSend("user.exchange","order.routingkey",orderRepository.findById(id).get());
+    public void getOrderByIdDemoRabbit(Long id) {
+//        System.out.println(orderRepository.findById(id).get());
+        Order order = orderRepository.findById(id).get();
+//        System.out.println(order.getOrderDate());
+        LocalDate orderDate = order.getOrderDate();
+        System.out.println(orderDate);
+        rabbitTemplate.convertAndSend("user.exchange", "order.routingkey", order);
+    }
+
+    public <T, D> T convertModel(D obj, Class<T> classT) {
+        ModelMapper modelMapper = new ModelMapper();
+        T obj1 = modelMapper.map(obj, classT);
+        return obj1;
+    }
+
+    public <T, D> List<T> convertListModel(List<D> objList, Class<T> classT) {
+        List<T> objResults = new ArrayList<T>();
+        for (D obj : objList) {
+            ModelMapper modelMapper = new ModelMapper();
+            T objResult = modelMapper.map(obj, classT);
+            objResults.add(objResult);
+        }
+        return objResults;
     }
 }
